requirements.txt
streamlit
pandas
google-generativeai
plotly
python-dotenv
fuzzywuzzy
python-Levenshtein